[C-5580] Remove legacy search page
