+++ packages/web/src/pages/search-page/RecentSearches.tsx
