+++ packages/web/src/pages/search-page/SearchCatalogTile.tsx
