+++ packages/web/src/pages/search-page/categories.ts
