--- packages/web/src/pages/search-page/components/desktop/SearchPageContent.d.ts
@@ -1,5 +1,0 @@
-import { SearchPageContentProps } from '../mobile/SearchPageContent'
-
-declare const SearchPageContent = (props: SearchPageContentProps) => JSX.Element
-
-export default SearchPageProvider
