--- packages/web/src/pages/search-page/components/desktop/SearchPageContent.module.css
@@ -1,92 +1,0 @@
-.searchResultsHeader {
-  justify-content: left;
-}
-
-.searchText {
-  margin-left: 32px;
-  color: var(--harmony-neutral);
-  font-size: var(--harmony-font-2xl);
-  line-height: 29px;
-  text-shadow: 0 2px 6px 0 var(--harmony-neutral);
-  font-weight: var(--harmony-font-regular);
-}
-
-.searchResultsSection {
-  text-align: left;
-  margin-top: 32px;
-}
-
-.searchResultsSection h2 {
-  color: var(--harmony-neutral);
-  font-size: var(--harmony-font-2xl);
-  font-weight: var(--harmony-font-bold);
-  line-height: 29px;
-  text-shadow: 0 2px 6px 0 var(--harmony-neutral);
-}
-
-.cardsContainer {
-  display: flex;
-  width: 100%;
-  justify-content: flex-start;
-  gap: var(--harmony-spacing-xl);
-}
-
-/* For mock data */
-.cardArt {
-  height: 136px;
-  width: 136px;
-  border: 2.27px solid #ffffff;
-  box-shadow: 0 1px 3px -2px rgba(133, 129, 153, 0.3);
-}
-
-.artistSearchResultsContainer {
-  margin-bottom: 64px;
-}
-
-.trackSearchResultsContainer {
-  margin-bottom: 64px;
-}
-
-.playlistSearchResultsContainer {
-  margin-bottom: 32px;
-}
-
-.albumSearchResultsContainer {
-  margin-bottom: 64px;
-}
-
-.spinner {
-  height: 40px;
-  width: 40px;
-  margin: 0 auto;
-}
-
-.spinner g path {
-  stroke: var(--harmony-secondary) !important;
-}
-
-.noResults {
-  text-align: center;
-  margin-top: 80px;
-}
-
-.noResults > svg {
-  height: 64px;
-  width: 64px;
-  margin-bottom: 8px;
-}
-
-.queryText {
-  font-size: var(--harmony-font-l);
-  font-weight: var(--harmony-font-medium);
-  line-height: 25px;
-  text-align: center;
-  text-shadow: 0 2px 6px 0 var(--harmony-neutral);
-}
-
-.helperText {
-  margin-top: 24px;
-  font-size: var(--harmony-font-m);
-  line-height: 25px;
-  text-shadow: 0 2px 6px 0 var(--harmony-neutral);
-}
