--- packages/web/src/pages/search-page/components/mobile/SearchPageContent.module.css
@@ -1,69 +1,0 @@
-.header {
-  border-bottom: 1px solid var(--harmony-n-100);
-}
-
-.lineupContainer {
-  margin-top: 12px;
-  padding: 0px 11px;
-  min-width: 0;
-  width: 100%;
-}
-
-.pageContainer {
-  --tab-height: 48px;
-  margin-top: var(--tab-height);
-  width: 100vw;
-}
-
-.tabContainer {
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-  margin-top: var(--header-padding-top);
-}
-
-.tabBar {
-  position: fixed;
-  top: 92px;
-  margin-top: 0px;
-  margin-top: env(safe-area-inset-top, 0px);
-  left: 0;
-  right: 0;
-  z-index: 1;
-}
-
-.spinner {
-  height: 40px;
-  width: 40px;
-  margin: 0 auto;
-}
-
-.spinner g path {
-  stroke: var(--harmony-secondary) !important;
-}
-
-.noResults {
-  color: var(--harmony-neutral);
-  font-size: var(--harmony-font-m);
-  font-weight: var(--harmony-font-medium);
-  max-width: 305px;
-  padding: 25px 35px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  min-width: 0;
-}
-
-.noResults svg {
-  height: 60px;
-  width: 60px;
-}
-
-.noResults span {
-  color: var(--harmony-n-400);
-  font-size: var(--harmony-font-l);
-}
-
-.noResults > * {
-  margin-bottom: 16px;
-}
