+++ packages/web/src/pages/search-page/types.ts
