+++ packages/web/src/pages/track-page/DesktopServerTrackPage.tsx
@@ -165,7 +165,7 @@ export const DesktopServerTrackPage = () => {
                   .split(',')
                   .filter((t) => t)
                   .map((tag) => (
-                    <Link key={tag} to={searchResultsPage(`#${tag}`)}>
+                    <Link key={tag} to={searchResultsPage('all', `#${tag}`)}>
                       <Tag>{tag}</Tag>
                     </Link>
                   ))}
